ya arregle generar rayos

+++ src/misc.h
@@ -14,5 +14,5 @@
 int main_kfold(int pca, int k, int alfa, int kknn);
 double ErrorCuadraticoMedio(Vector &b, Vector &bcalculado);
 int grabarPPM8Bits(std::string nombre_archivo, int filas, int columnas, Vector I);
-vector<pair<pair<int, int>, double>> generarRayosVertices(int tam,int cantRayos);
+vector<pair<pair<double, double>, double>> generarRayosVertices(int tam,int cantRayos);
 #endif
